WIP for my `PgPtr` idea.  Nothing here works right now.  Things don't even compile.
I'm only pushint this as a request via https://github.com/intellij-rust/intellij-rust/issues/6522#issuecomment-746852495

+++ pgx-pg-sys/src/pg12_oids.rs
@@ -164,7 +164,7 @@ pub enum PgBuiltInOids {
 }
 impl PgBuiltInOids {
     pub fn from(oid: crate::Oid) -> Option<PgBuiltInOids> {
-        match oid {
+        match oid as i32 {
             crate::HEAP_TABLE_AM_HANDLER_OID => {
                 Some(crate::PgBuiltInOids::HEAP_TABLE_AM_HANDLER_OID)
             }
